backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL驱动
# python-multipart==0.0.6  # 文件上传支持

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI集成
google-generativeai==0.3.2
openai==1.3.0
langchain==0.0.340
langchain-google-genai==0.0.2

# 环境变量管理
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# 日志系统
structlog==23.2.0
python-json-logger==2.0.7

# 生产环境
gunicorn==21.2.0
python-multipart==0.0.6